+++ src/pflotran/translator_mixed_fluid_OWG.F90
@@ -296,7 +296,7 @@
   integer icri,ichange, itable, ierr 
 
   PetscScalar, pointer :: xx_p(:), yy_p(:),iphase_p(:)
-  integer*4 :: n,n0,index,ipr
+  integer :: n,n0,index,ipr
   integer :: iipha,i 
   real*8 :: p2,p,tmp,t, xla
   real*8 :: dg,dddt,dddp,fg,dfgdp,dfgdt,eng,hg,dhdt,dhdp,visg,dvdt,dvdp
